pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-restx>=1.3.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "kafka-python>=2.0.6",
    "prometheus-client>=0.21.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "werkzeug>=3.1.3",
]
